kb-forall: Rename 'skip' to 'exact'
We test if commit is PICKED, the 'skip' was not meant the test is not
performed, but that the commit is exactly there with the same sha1.
The commit would be reported by `git br --contains`, but 'contains' is
the word with the same length as 'unpicked' so it would be harder to
notice. 'identical' is  too long and just 1 char different (as
the 'contain').

'exact' should be interpreted as the comment is exactly the same.

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>